% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMatrix.R
\name{corrMatrix}
\alias{corrMatrix}
\title{Correlation Matrix of Numeric Variables in a Year}
\usage{
corrMatrix(year)
}
\arguments{
\item{year}{year selected. Must be between 1950 and 2017 (included).}
}
\description{
This function returns the correlation matrix of all numeric variables in a selected year. Only complete pairs of observations are considered.
}
\examples{
corrMatrix(2015)
               Age           G          GS          MP         PER         TS.
Age    1.000000000  0.12608101  0.02921544  0.08427518  0.08438978  0.11508346
...
}
\keyword{correlation,}
\keyword{matrix,}
\keyword{numeric}
