% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KobePER.R
\name{KobePER}
\alias{KobePER}
\title{Kobe Bryant's Player Efficiency Rate in Year}
\usage{
KobePER(year)
}
\arguments{
\item{year}{year selected. Must be between 1997 and 2016 (included).}
}
\description{
This function displays the Player Efficiency Rate (PER) of Kobe Bryant in the year selected.
}
\examples{
KobePER(2015)
17.6
}
\keyword{Bryant}
\keyword{Kobe,}
\keyword{PER,}
