% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldestPlayer.R
\name{oldestPlayer}
\alias{oldestPlayer}
\title{Oldest Player in a Year}
\usage{
oldestPlayer(year)
}
\arguments{
\item{year}{year selected. Must be between 1950 and 2017 (included).}
}
\description{
This function shows the name of the oldest player in the year selected. In case of multiple players, the first name in alphabetical order will be selected.
}
\examples{
oldestPlayer(2015)
"Andre Miller"
}
\keyword{oldest,}
\keyword{player}
